Consolidate to two binaries, add SSDP discovery and USB disk to CLI/TUI
- Extract internal/disk/ package from cmd/diskutil/ for reusable disk management
- Add discover and disk subcommands to panago-cli
- Replace TUI IP form with SSDP auto-discovery screen + manual IP fallback
- Implement TUI disk menu with listing, confirmation, and progress
- Add ExtractIP() and DiscoverPanasonic() helpers to internal/upnp/
- Remove standalone binaries: cmd/{discover,diskutil,firmware,cramfs,shell,keydump}/
- Remove dead code: unused packages (punch, shell, config, logging),
  unused functions (findNextPartition, ExtractFMA5, EncodeSimple, findPatchFolders)
- Update Makefile, .goreleaser.yaml, README.md for two-binary setup
- Run go mod tidy to clean unused dependencies

+++ pkg/cramfs/build.go
@@ -124,7 +124,7 @@ func (b *Builder) collectEntries(dirPath string) ([]entryInfo, error) {
 		entry := entryInfo{
 			name: name,
 			path: fullPath,
-			mode: uint16(info.Mode().Perm()) | uint16(info.Mode()&fs.ModeType)>>16,
+			mode: uint16(info.Mode().Perm()) | uint16(uint32(info.Mode()&fs.ModeType)>>16),
 			uid:  0, // Could extract from syscall if needed
 			gid:  0,
 		}
